CBRL: extract restore orchestration into CbrlRestore (cbrl package)
Move the whole restore pipeline out of the integration-test harness into a new
public CbrlRestore.restore(): drive PREPARED-record recovery against the
reloaded coordinator backup, explode the committed redo, shuffle, replay via
RecordApplier (read-and-compute only), and write the result back to the restore
tables. The write-back now lives in the cbrl package, not the test.

CoordinatorBackupRow moves to the cbrl package so the IT (which builds it) and
CbrlRestore (which consumes it) share it. The IT now calls only restore() and
asserts on the restored tables.

CbrlRestore is the package's only public class; every other cbrl type and its
members are package-private, since the unit and integration test source sets
share the package. Record this contract as section 4.7 in the PoC plan.
